Reorganize folders and files.
Remove scripts to build this extension for PHP 7.4 and below.

+++ lib/biomes/BiomeClimate.cpp
@@ -1,7 +1,7 @@
 #include <map>
-#include <lib/objects/Biome.h>
-#include <lib/pocketmine/BiomeList.h>
-#include <lib/vanilla/octaves/SimplexOctaveGenerator.h>
+#include <lib/biomes/BiomeClimate.h>
+#include <lib/objects/constants/BiomeList.h>
+#include <lib/noise/octaves/SimplexOctaveGenerator.h>
 
 namespace Biome {
 
@@ -22,7 +22,7 @@ double getVariatedTemperature(uint_fast8_t biome, int_fast64_t x, int_fast32_t y
 
   temp = get(biome).temperature;
   if (y > 64) {
-    variation = noise_gen->noise(x, z, 0, 0.5, 2.0, false) * 4.0;
+    variation = noise_gen->noise(static_cast<double>(x), static_cast<double>(z), 0, 0.5, 2.0, false) * 4.0;
 
     return temp - (variation + (float) (y - 64)) * 0.05 / 30.0;
   }
